Deploy: Version 2.0

+++ MeneMarket/Services/Processings/Files/FileProcessingService.cs
@@ -1,4 +1,5 @@
 ﻿using MeneMarket.Services.Foundations.Files;
+using SixLabors.ImageSharp;
 
 namespace MeneMarket.Services.Processings.Files
 {
@@ -13,7 +14,7 @@ namespace MeneMarket.Services.Processings.Files
 
         public async ValueTask<string> UploadFileAsync(
             MemoryStream memoryStream,
-            string fileName)
+            string fileName, int maxWidth, int maxHeight)
         {
             var uploadsFolder =
                 Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "imageFiles");
@@ -23,21 +24,30 @@ namespace MeneMarket.Services.Processings.Files
                 Directory.CreateDirectory(uploadsFolder);
             }
 
-            return await this.fileService.UploadFileAsync(memoryStream, fileName, uploadsFolder);
+            return await this.fileService.UploadFileAsync(memoryStream, fileName, uploadsFolder, maxWidth, maxHeight);
         }
 
-        public string DeleteImageFile(string imageName)
+        public string DeleteImageFile(List<string> filePaths)
         {
-            string filePath = Path.Combine("wwwroot\\imageFiles", imageName);
-
-            if (System.IO.File.Exists(filePath))
+            if (filePaths.Count != 0)
             {
-                this.fileService.RemoveImageFile(filePath);
-                return "Image deleted successfully.";
+                foreach (var filePath in filePaths)
+                {
+                    if (System.IO.File.Exists(filePath))
+                    {
+                        this.fileService.RemoveImageFile(filePath);
+                    }
+                    else
+                    {
+                        return "Image not found.";
+                    }
+                }
+
+                return "Images deleted successfully.";
             }
             else
             {
-                return "Image not found.";
+                return "file Paths is null";
             }
         }
     }
